+++ CHANGELOG.md
@@ -26,7 +26,8 @@ The sticker database was removed in this release. myMPD fetches now all stickers
 - Feat: Add exclude and include filter expression for jukebox mode #1098
 - Feat: Add min. duration option for jukebox mode #1098
 - Feat: Respect search order for searches added to queue or playlist #1078
-- Feat: add support for descending sort for smart playlists and playlists
+- Feat: Add support for descending sort for smart playlists and playlists
+- Feat: Improve jukebox album mode
 - Upd: Use more standard cmake build types #1119
 - Upd: rework timer implementation
 - Upd: some websocket connection improvements
@@ -41,7 +42,7 @@ This is a small bug fix release.
 ### Changelog
 
 - Fix: Smart playlist option disappears after it is turned off in settings #1133
-- Fix: Fill jukebox queue respects function should respect current length
+- Fix: Fill jukebox queue function should respect current length
 
 ***